feat: implement global search and filter functionality in MainLayout and Home components
