+++ src/pages/Home.jsx
@@ -1,13 +1,16 @@
 import { useState, useEffect } from 'react'
+import { useOutletContext, Link } from 'react-router-dom';
 import { heatPumpService } from '../services/heatPumpService';
 import pumpLogo from '../assets/logo-placeholder.png'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faXmark, faCircleCheck, faTriangleExclamation, faCircleQuestion, faLocationDot, faTemperatureHalf, faGauge, faBolt, faFilePen } from '@fortawesome/free-solid-svg-icons';
-import { Link } from 'react-router-dom';
+import { faCircleCheck, faTriangleExclamation, faCircleQuestion, faLocationDot, faTemperatureHalf, faGauge, faBolt, faFilePen } from '@fortawesome/free-solid-svg-icons';
 
 import { format } from 'date-fns';
 
 function Home() {
+    // Get global filter state from MainLayout
+    const { search, filters } = useOutletContext();
+    
     const formatDate = (dateString) => {
         if (!dateString) return '-';
         try {
@@ -23,15 +26,7 @@ function Home() {
     const [loading, setLoading] = useState(false);
     const [error, setError] = useState(null);
     const [data, setData] = useState([]);
-
-    const [search, setSearch] = useState(null);
-    const [filters, setFilters] = useState({
-        fabrikant: null,
-        bedrijf: null,
-        merk: null
-    });
     const [pompen, setPompen] = useState([]);
-    const [inputValue, setInputValue] = useState(""); // Controlled input state
 
     useEffect(() => {
         const fetchData = async () => {
@@ -100,68 +95,6 @@ function Home() {
         setPompen(filterHeatPumps(data, filters, search));
     }, [search, filters, data]);
 
-    const handleSearch = (e) => {
-        const val = e.target.value;
-        setInputValue(val);
-        setSearch(val === '' ? null : val.toLowerCase());
-    }
-
-    const resetSearch = () => {
-        setInputValue("");
-        setSearch(null);
-    }
-
-    const toggleFilter = () => {
-        let sidebarFilter = document.getElementById('sidebarFilter');
-        if (sidebarFilter.classList.contains('hidden')) {
-            sidebarFilter.classList.remove('hidden');
-        } else {
-            sidebarFilter.classList.add('hidden');
-        }
-    }
-
-    const toggleDropdown = (dropdownId) => {
-        let dropdownFilter = document.getElementById(dropdownId);
-        if (dropdownFilter.classList.contains('hidden')) {
-            dropdownFilter.classList.remove('hidden');
-        } else {
-            dropdownFilter.classList.add('hidden');
-        }
-    }
-
-    const selectFilter = (filter) => {
-        const key = Object.keys(filter)[0];
-        const newFilters = { ...filters };
-
-        if (filters[key] !== null) {
-            if (filters[key] === filter[key]) {
-                newFilters[key] = null;
-            } else {
-                newFilters[key] = filter[key];
-            }
-        } else {
-            newFilters[key] = filter[key];
-        }
-
-        setFilters(newFilters);
-    }
-
-    const isSelected = (key, fltrSubject) => {
-        if (filters[key] && filters[key].includes(fltrSubject)) {
-            return true;
-        } else {
-            return false;
-        }
-    }
-
-    const resetFilter = () => {
-        setFilters({
-            fabrikant: null,
-            bedrijf: null,
-            merk: null
-        });
-    }
-
     const pompStatus = (status) => {
         if (status == 200) {
             return {
@@ -186,90 +119,8 @@ function Home() {
         }
     }
 
-    // Helper to get unique values for filters
-    const getUniqueValues = (key) => {
-        if (!data) return [];
-        const values = data.map(item => item[key]).filter(v => v !== null && v !== undefined && v !== '');
-        return [...new Set(values)].sort(); // Sort alphabetically
-    };
-
     return (
         <>
-            <nav className="fixed top-0 left-64 right-0 z-40 bg-white">
-                <div className='flex items-center justify-between w-full p-3'>
-                    <div className="flex items-center justify-around w-full">
-                        <input 
-                            className="w-full p-2 text-sm text-gray-900 border border-gray-300 rounded-lg mr-2 bg-gray-50 focus:ring-blue-500 focus:border-blue-500" 
-                            onChange={handleSearch} 
-                            value={inputValue} 
-                            type="text" 
-                            placeholder="Voer postcode in..." 
-                        />
-                        {search != null && <FontAwesomeIcon className="cursor-pointer" onClick={resetSearch} icon={faXmark} />}
-                    </div>
-                    <div>
-                        <span className='p-2 cursor-pointer' onClick={toggleFilter}>Filter</span>
-                    </div>
-                </div>
-            </nav>
-            <aside id='sidebarFilter' className="fixed top-0 right-0 z-50 w-64 h-screen bg-white border-r border-gray-200 hidden">
-                <div className="h-full overflow-y-auto">
-                    <div className="p-3 mt-2">
-                        <span className='p-2 cursor-pointer' onClick={toggleFilter}>Sluiten</span>
-                    </div>
-                    <hr />
-                    <div className="p-3">
-                        <ul className="p-2">
-                            <li>
-                                <span className="rounded-lg ml-3 cursor-pointer select-none" onClick={() => toggleDropdown('dropdownFabrikantFilter')}>Fabrikant</span>
-
-                                <div id="dropdownFabrikantFilter" className="z-10 hidden bg-white divide-y divide-gray-100 rounded-lg w-44">
-                                    <ul className="text-sm text-gray-700 border rounded-lg my-2">
-                                        <li onClick={() => selectFilter({ fabrikant: 'Fabrikant'})}>
-                                            <span className="block px-4 py-2 rounded-lg hover:bg-gray-100">Fabrikant Filter</span>
-                                        </li>
-                                    </ul>
-                                </div>
-                            </li>
-                            <li>
-                                <span className="rounded-lg ml-3 cursor-pointer select-none" onClick={() => toggleDropdown('dropdownBedrijfFilter')}>Bedrijf</span>
-
-                                <div id="dropdownBedrijfFilter" className="z-10 hidden bg-white divide-y divide-gray-100 rounded-lg w-44">
-                                    <ul className="text-sm text-gray-700 border rounded-lg my-2">
-                                        <li onClick={() => selectFilter({ bedrijf: 'Intergas'})}>
-                                            <span className="block px-4 py-2 rounded-lg hover:bg-gray-100">Intergas</span>
-                                        </li>
-                                        <li onClick={() => selectFilter({ bedrijf: 'Remeha'})}>
-                                            <span className="block px-4 py-2 rounded-lg hover:bg-gray-100">Remeha</span>
-                                        </li>
-                                        <li onClick={() => selectFilter({ bedrijf: 'Bosch'})}>
-                                            <span className="block px-4 py-2 rounded-lg hover:bg-gray-100">Bosch</span>
-                                        </li>
-                                        <li onClick={() => selectFilter({ bedrijf: 'Vaillant'})}>
-                                            <span className="block px-4 py-2 rounded-lg hover:bg-gray-100">Vaillant</span>
-                                        </li>
-                                    </ul>
-                                </div>
-                            </li>
-                            <li>
-                                <span className="rounded-lg ml-3 cursor-pointer select-none" onClick={() => toggleDropdown('dropdownMerkFilter')}>Merk/Type</span>
-
-                                <div id="dropdownMerkFilter" className="z-10 hidden bg-white divide-y divide-gray-100 rounded-lg w-44">
-                                    <ul className="text-sm text-gray-700 border rounded-lg my-2">
-                                        <li onClick={() => selectFilter({ merk: 'Merk'})}>
-                                            <span className="block px-4 py-2 rounded-lg hover:bg-gray-100">Merk</span>
-                                        </li>
-                                    </ul>
-                                </div>
-                            </li>
-                        </ul>
-                    </div>
-                    <hr />
-                    <div className="p-3 mt-2">
-                        <span className='p-2 cursor-pointer hover:text-red-500' onClick={resetFilter}>Filters Wissen</span>
-                    </div>
-                </div>
-            </aside>
             <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4 min-[2500px]:grid-cols-5 gap-6">
                 {error && (
                     <div className="col-span-full p-4 text-red-700 bg-red-100 border border-red-400 rounded w-full">
